refactor: hooks 관심사에 따른 분리
