+++ src/app/group/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useTheme } from '@/hooks/useTheme';
+import { useTheme } from '@/hooks/ui/useTheme';
 import { spacing } from '@/styles/design-system';
 import GroupSidebar from '@/components/group/GroupSidebar';
