+++ src/app/group/team/[id]/page.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useTheme } from '@/hooks/useTheme';
+import { useTheme } from '@/hooks/ui/useTheme';
 import { Card, CardContent } from '@/shadcn/ui/card';
 import { spacing } from '@/styles/design-system';
 import { useParams } from 'next/navigation';
