+++ src/components/leaderboard/CategoryFilter.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { useTheme } from '@/hooks/useTheme';
+import { useTheme } from '@/hooks/ui/useTheme';
 import { categoryColors } from '@/styles';
 import { FONT_SIZES } from '@/styles/font-sizes';
