+++ src/hooks/data/useLeaderboardInfiniteScroll.ts
@@ -6,7 +6,7 @@ import {
   getKSTMonthlyDateString,
   getKSTWeeklyDateString,
 } from '@/utils/timezone';
-import { useInfiniteScroll } from './useInfiniteScroll';
+import { useInfiniteScroll } from '../common/useInfiniteScroll';
 import { INFINITE_SCROLL_CONFIG } from '@/config/constants/infinite-scroll';
 
 // User 타입은 userStore에서 import
